(Combine multiple admin tickets for currency mismatch event for the same stockcode)
We were caching tickets written to Vimaly, but not updating the cache when creating and updating new tickets, leading to duplication when the same issue occurred repeatedly in the same session (as above).
Since vimaly logging is an exceptional event, we don't need a cache, so remove it and treat Vimaly as the single source of truth.